refactor(files): improve render of file upload component
